finishing up

+++ controllers/ProductController.js
@@ -8,15 +8,29 @@ const {
   response,
   pushUpdate,
   pullUpdate,
+  read,
+  tryCatchError,
 } = require("../utils");
+const { readTotal } = require("../utils/crud");
 
 const createProduct = async (req, res) => {
   const fullUrl = req.protocol + "://" + req.get("host");
-  if (!req.file) return response(res, 404, { message: "File Not Uploaded!" });
-  const filePath = req.file.path.replace(/\\/g, "/");
+  const files = req.files;
+  console.log(files);
+  if (!files || files.length === 0)
+    return response(res, 404, { message: "File Not Uploaded!" });
+  const filePaths = files.map((file) => {
+    return `${fullUrl}/${file.path.replace(/\\/g, "/")}`;
+  });
+  console.log(req.body.colors);
   const product = await create(
     res,
-    { ...req.body, retailer: req.id, image: `${fullUrl}/${filePath}` },
+    {
+      ...req.body,
+      retailer: req.id,
+      images: filePaths,
+      colors: req.body.colors ? req.body.colors.split(",") : [],
+    },
     Product,
     false
   );
@@ -41,24 +55,67 @@ const createProduct = async (req, res) => {
 const getProduct = async (req, res) => {
   await readOne(res, { id: req.headers["id"] }, Product);
 };
+const getTotalProducts = async (req, res) => {
+  await readTotal(res, Product);
+};
+const getTotalProductsByRetailer = async (req, res) => {
+  if (!req.headers["id"])
+    return response(res, 409, { message: "Retailer id not received" });
+  await readTotal(res, Product, { retailer: req.headers["id"] });
+};
+const getProducts = async (req, res) => {
+  await read(res, {}, Product, ["category", "subCategory"]);
+};
 
 const getProductsByRetailer = async (req, res) => {
-  await readBy(res, { retailer: req.id }, Product);
+  await readBy(res, { retailer: req.headers["id"] }, Product, [
+    "category",
+    "subCategory",
+  ]);
+};
+const getProductsByCategory = async (req, res) => {
+  await readBy(res, { category: req.headers["id"] }, Product, [
+    "category",
+    "subCategory",
+  ]);
+};
+const getFeaturedProducts = async (req, res) => {
+  await readBy(res, { featured: true }, Product);
+};
+const getPopularProducts = async (req, res) => {
+  try {
+    const topProducts = await Product.find({}).sort({ orders: -1 }).limit(9);
+    return response(res, 200, { items: topProducts });
+  } catch (error) {
+    return tryCatchError(res, error);
+  }
+};
+const getProductsBySubCategory = async (req, res) => {
+  await readBy(res, { subCategory: req.headers["id"] }, Product, [
+    "category",
+    "subCategory",
+  ]);
 };
 const updateProduct = async (req, res) => {
   const fullUrl = req.protocol + "://" + req.get("host");
+  const files = req.files;
+  console.log("these are files", files);
   let query;
-  if (req.file) {
-    const filePath = req.file.path.replace(/\\/g, "/");
-
-    query = {
-      ...req.body,
-      image: `${fullUrl}/${filePath}`,
-    };
-  } else query = req.body;
+  if (!files || files.length === 0)
+    return response(res, 404, { message: "File Not Uploaded!" });
+  const filePaths = files.map((file) => {
+    if (file.path.startsWith(fullUrl))
+      return `${file.path.replace(/\\/g, "/")}`;
+    else return `${fullUrl}/${file.path.replace(/\\/g, "/")}`;
+  });
+  query = {
+    ...req.body,
+    colors: req.body.colors.split(","),
+    images: filePaths,
+  };
   const product = await update(res, query, Product, req.headers["id"], false);
   if (req.body.category) {
-    await pushUpdate(
+    await pullUpdate(
       res,
       product._id,
       Category,
@@ -66,7 +123,7 @@ const updateProduct = async (req, res) => {
       "products",
       false
     );
-    await pullUpdate(
+    const resp = await pushUpdate(
       res,
       product._id,
       Category,
@@ -76,7 +133,7 @@ const updateProduct = async (req, res) => {
     );
   }
   if (req.body.subCategory) {
-    await pushUpdate(
+    await pullUpdate(
       res,
       product._id,
       SubCategory,
@@ -84,7 +141,7 @@ const updateProduct = async (req, res) => {
       "products",
       false
     );
-    await pullUpdate(
+    await pushUpdate(
       res,
       product._id,
       SubCategory,
@@ -95,16 +152,36 @@ const updateProduct = async (req, res) => {
   }
   return response(res, 200, { message: "Data Updated!", product });
 };
+const featureUnFeatureProduct = async (req, res) => {
+  try {
+    const productId = req.headers["id"];
+    if (!productId)
+      return response(res, 400, { message: "Product Id Not Received!" });
+    const product = await Product.findById(productId);
+    if (!product)
+      return response(res, 404, {
+        message: "Product with this Id not found!",
+      });
+    await Product.findByIdAndUpdate(productId, { featured: !product.featured });
+    return response(res, 200, {
+      message: `Product ${product.featured ? "UnFeatured" : "Featured"} `,
+    });
+  } catch (error) {
+    tryCatchError(res, error);
+  }
+};
 const deleteProduct = async (req, res) => {
-  await pullUpdate(
+  const subCatResp = await pullUpdate(
     res,
     req.headers["id"],
     SubCategory,
-    req.headers["subCategory"],
+    req.headers["subcategory"],
     "products",
     false
   );
-  await pullUpdate(
+  if (subCatResp.success === false)
+    return response(res, subCatResp.status, { message: subCatResp.message });
+  const catResp = await pullUpdate(
     res,
     req.headers["id"],
     Category,
@@ -112,6 +189,9 @@ const deleteProduct = async (req, res) => {
     "products",
     false
   );
+  if (catResp.success === false)
+    return response(res, catResp.status, { message: catResp.message });
+
   await deleteItem(res, { id: req.headers["id"] }, Product);
 };
 module.exports = {
@@ -120,4 +200,12 @@ module.exports = {
   updateProduct,
   getProductsByRetailer,
   deleteProduct,
+  getProductsByCategory,
+  getProductsBySubCategory,
+  getProducts,
+  getTotalProducts,
+  getTotalProductsByRetailer,
+  featureUnFeatureProduct,
+  getFeaturedProducts,
+  getPopularProducts,
 };
